+++ src/popo/reservation/place/reserve.place.controller.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   Body,
   Controller,
   Delete,
@@ -27,6 +28,7 @@ import { RolesGuard } from '../../../auth/authroization/roles.guard';
 import { PlaceService } from '../../place/place.service';
 import {JwtPayload} from "../../../auth/strategies/jwt.payload";
 import { ReservePlace } from './reserve.place.entity';
+import * as moment from 'moment-timezone';
 
 @ApiTags('Place Reservation')
 @Controller('reservation-place')
@@ -281,13 +283,21 @@ export class ReservePlaceController {
   @UseGuards(JwtAuthGuard)
   async delete(@Param('uuid') uuid: string, @Req() req) {
     const reservation = await this.reservePlaceService.findOneByUuidOrFail(uuid);
-    const user = req.user;
+    const user = req.user as JwtPayload;
 
     if (user.userType == UserType.admin || user.userType == UserType.staff) {
       await this.reservePlaceService.remove(uuid);
     } else {
       if (reservation.booker_id == user.uuid) {
-        await this.reservePlaceService.remove(uuid);
+        // if the reservation is in the past, deny delete
+        const reservation_start_time = reservation.date + reservation.start_time;
+        const current_time = moment().tz('Asia/Seoul').format('YYYYMMDDHHmm');
+
+        if (reservation_start_time < current_time) {
+          throw new BadRequestException('Cannot delete past reservation');
+        } else {
+          await this.reservePlaceService.remove(uuid);
+        }
       } else {
         throw new UnauthorizedException('Unauthorized delete action');
       }
